+++ src/components/Checkbox/view/alpha/Checkbox.tsx
@@ -20,6 +20,7 @@ export const Checkbox: CheckboxType = ({
     id,
     isChecked,
     className,
+    size = 'small',
     ...props
 }) => {
     const springRef = useRef(null);
@@ -47,11 +48,13 @@ export const Checkbox: CheckboxType = ({
             isDisabled={isDisabled}
             className={className}
             htmlFor={id}
+            size={size}
         >
             {horizontalPosition === 'right' && children && (
                 <StyledCheckboxSpan
                     position={horizontalPosition}
                     isDisabled={isDisabled}
+                    size={size}
                 >
                     {children}
                 </StyledCheckboxSpan>
@@ -88,6 +91,7 @@ export const Checkbox: CheckboxType = ({
                 <StyledCheckboxSpan
                     position={horizontalPosition}
                     isDisabled={isDisabled}
+                    size={size}
                 >
                     {children}
                 </StyledCheckboxSpan>