+++ src/components/Input/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useInputChange';
+export * from './useInputClearable';